Feat : 사용자 확인 후 api 진행 및 레슨 생성자 아이디 생성 과정 변경

+++ src/main/java/com/example/nanuri/domain/lesson/LessonRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface LessonRepository extends JpaRepository<Lesson,Integer> {
+public interface LessonRepository extends JpaRepository<Lesson,Long> {
 
     @Modifying
     @Query(
